Divide class into packages
