--- src/main/java/Command.java
@@ -1,5 +1,0 @@
-public abstract class Command {
-    public abstract void execute(TaskList taskList, 
-                                 Storage storage, Ui ui);
-    public abstract boolean isRunning();
-}
