+++ src/main/java/roofus/Storage.java
@@ -1,3 +1,5 @@
+package roofus;
+
 import java.io.*;
 
 import java.util.List;
