+++ src/main/java/roofus/command/ByeCommand.java
@@ -1,8 +1,14 @@
+package roofus.command;
+
+import roofus.Storage;
+import roofus.TaskList;
+import roofus.Ui;
+
 import java.io.IOException;
 
 public class ByeCommand extends Command {
     @Override 
-    public void execute(TaskList taskList, 
+    public void execute(TaskList taskList,
                         Storage storage, Ui ui) {
         ui.signOff();
         try {
