+++ src/main/java/roofus/command/ListCommand.java
@@ -1,6 +1,12 @@
+package roofus.command;
+
+import roofus.Storage;
+import roofus.TaskList;
+import roofus.Ui;
+
 public class ListCommand extends Command {
     @Override
-    public void execute(TaskList taskList, 
+    public void execute(TaskList taskList,
                         Storage storage, Ui ui) {
         ui.list(taskList);
     }
