+++ src/main/java/roofus/task/Event.java
@@ -1,6 +1,6 @@
+package roofus.task;
+
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.time.temporal.ChronoUnit;
 
 public class Event extends Task {
     LocalDate start;
