+++ src/main/java/roofus/task/ToDo.java
@@ -1,5 +1,6 @@
-public class ToDo extends Task {
+package roofus.task;
 
+public class ToDo extends Task {
     public ToDo(String description) {
         super(description);
     }
